refactor: 패키지 오타 수정
domian > domain
