+++ feature/category/src/main/java/com/chan/category/di/CategoryRepositoryModule.kt
@@ -2,8 +2,8 @@ package com.chan.category.di
 
 import com.chan.category.data.repositoryImpl.CategoryDetailRepositoryImpl
 import com.chan.category.data.repositoryImpl.CategoryRepositoryImpl
-import com.chan.category.domian.CategoryDetailRepository
-import com.chan.category.domian.CategoryRepository
+import com.chan.category.domain.CategoryDetailRepository
+import com.chan.category.domain.CategoryRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
