+++ feature/category/src/main/java/com/chan/category/ui/mapper/CategoryMappers.kt
@@ -1,6 +1,6 @@
 package com.chan.category.ui.mapper
 
-import com.chan.category.domian.vo.CategoryVO
+import com.chan.category.domain.vo.CategoryVO
 import com.chan.category.ui.model.CategoryModel
 
 fun CategoryVO.toPresentation(): CategoryModel {
